Added delete expense validation

+++ delete_expense.php
@@ -1,7 +1,18 @@
 <?php
 include "db.php";
+if($_SERVER["REQUEST_METHOD"]!="GET"){
+    die("Invalid Request");
+}
 
+if(!isset($_GET['id'])){
+    die ("Invalid Id");
+}
 $id=$_GET['id'];
+if($id<=0){
+    die ("Invalid Id");
+}
+
+
 $sql="DELETE FROM expenses WHERE id=?";
 
 $stmt=mysqli_prepare($con,$sql);
@@ -16,6 +27,6 @@ if($result){
     exit;
 }
 else{
-    echo "Error:".mysqli_error($result);
+    echo "Error:".mysqli_stmt_error($stmt);
 }
 ?>
